Update AdobeReader-Win.download.recipe
These changes allowed this download recipe to function for me.  The recipe was finding a version match but I was getting a variable undefined for display_version.  When I explicitly defined the display_version variable using result_output_var_name as part of the input variables for the URLTextSearcher processor the display_version variable was no longer coming up as undefined.  Also, I removed the bes_filepath key from both sections where the BESRelevanceProvider processor is called because the QnA path is already defined in the python script.